fix(measures): discover architecture.md via git, not a recursive repo walk
The architecture-drift health check walked the whole repo tree with
readdir to find architecture.md files, descending into gitignored runtime
data — notably infra/perf-stack/storage/ (Grafana-Tempo WAL), whose
directories are created and deleted at runtime. A scandir could therefore
ENOENT mid-walk and crash the check, breaking `npm run test:local` (the
pre-push gate) on a clean tree.

Replace the walk with `git ls-files --cached --others --exclude-standard
-- '*architecture.md'`. Git prunes ignored directories before it stats
their contents, so ephemeral/runtime trees are never touched — this
eliminates the failure mode at its source rather than catching ENOENT,
and matches the check's existing intent (it already skipped gitignored
architecture.md). Coverage is unchanged: tracked plus untracked-not-ignored
files, minus the same __tests__/brain/etc. excludes.

The 592-line implementation lived inside the .test.ts file and had grown
past the 500-line limit, so extract it into a focused module pair under
coupling/architecture-drift/:
  - architecture-contract.ts        markdown/Mermaid -> DiagramSpec
  - validate-architecture-drift.ts  discovery + drift validation
  - architecture-drift.test.ts      black-box tests
Grouping into a subfolder keeps coupling/ at the 15-child fanout limit.

Add a black-box regression test: discovery must exclude an architecture.md
under a gitignored runtime subtree and never scan its churning children.
