+++ packages/measures/src/checks/tier_1/structure/architecture-drift.ts
@@ -4,12 +4,12 @@ export const check: CheckDef = {
     id: 'architecture-drift',
     name: 'Architecture Drift',
     category: 'Static',
-    display: 'vitest run packages/measures/src/health/coupling/architecture-drift.test.ts',
+    display: 'vitest run packages/measures/src/health/coupling/architecture-drift/architecture-drift.test.ts',
     args: (jsonOut) => [
         'npx',
         'vitest',
         'run',
-        'packages/measures/src/health/coupling/architecture-drift.test.ts',
+        'packages/measures/src/health/coupling/architecture-drift/architecture-drift.test.ts',
         ...checkArgs.vitestJsonArgs(jsonOut),
     ],
     parser: 'vitest',
